+++ src/components/Todo/TodoList.js
@@ -1,14 +1,6 @@
 import PropTypes from 'prop-types';
 import { TodoItem } from './TodoItem';
 
-const styles = {
-  ul: {
-    listStyle: 'none',
-    margin: 0,
-    padding: 0,
-  },
-};
-
 export function TodoList(props) {
   return (
     <ul style={styles.ul}>
@@ -29,4 +21,12 @@ export function TodoList(props) {
 TodoList.propTypes = {
   todos: PropTypes.arrayOf(PropTypes.object).isRequired,
   onToggle: PropTypes.func.isRequired,
+};
+
+const styles = {
+  ul: {
+    listStyle: 'none',
+    margin: 0,
+    padding: 0,
+  },
 };
