--- src/context.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const Context = React.createContext()
-
-export default Context
